refactor: drive Error Prone to zero and hold nine checks at ERROR (#624)
* refactor: drive Error Prone to zero and hold nine checks at ERROR

Every build carried a standing backlog of Error Prone findings, so
nothing in it was read. Among them was a real defect:
InterfaceSnapshotPoller.refreshRegistrations() does the same
iterate-and-remove over registrations that tick() does, and never got
the annotation and justification tick() carries. Error Prone had been
pointing at it on every build since the loop was written.

The backlog was also under-reported. GitHub caps warning annotations at
ten per job, so the surface a maintainer reads had been showing ten
while the log held more. Counting every check rather than the five the
annotations named gives 16 in src/main, and compiling the test tree as
well gives 20. Recorded in the pom comment, since it is the reason the
backlog was invisible rather than merely ignored.

All twenty sites are at zero, each by the fix its cause called for:

- registrations is declared ConcurrentHashMap, so the contract three
  removal sites depend on is stated by the type instead of repeated in
  prose. tick()'s suppression became dead and is deleted.
- Three deliberate identity comparisons carry @SuppressWarnings naming
  the check, so the decision is greppable rather than implied by a
  comment. Two compare types with real value equality (Lombok
  @EqualsAndHashCode, a record); the third compares a lambda, where ==
  and .equals coincide and the annotation records intent.
- Two instanceof tests bind their pattern, dropping a cast each.
- Four borrowed parenthesis pairs deleted, leaving the concatenated
  form where they are load-bearing.
- One private method with no callers since #530 deleted.
- A javadoc stranded on the wrong test moved back to the test it
  documents, which is the one whose least obvious assertion it explains.

The nine checks then join the three already at ERROR, and each was
proven to gate: reintroduce the construct, build must fail at
compilation naming that check on that file. The first attempt at
PatternMatchingInstanceof returned inconclusive, not killed, because
the mutation left a dangling symbol and the build failed before the
check spoke.

LongDoubleConversion is fixed with an explicit cast rather than by
narrowing usable() to take a long. The narrower signature removed all
three warnings at the root and broke SamplingIntervalBoundaryTest,
which reflectively asserts every rate guard rejects 0.0, -1.0, 0.5, NaN
and +Inf — three of which a long parameter cannot express. That is a
design change to a sampling-correctness path, not a warning fix.

make jar: BUILD SUCCESS, 0 Error Prone findings across main and test,
Tests run: 2972, Failures: 0, Errors: 0, Skipped: 12. No *IT class runs
in that target, so this says nothing about the integration tests.


* chore: re-trigger CI

Riptide Build and Lint Actions sat queued for over an hour on the
previous commit with no runner assigned, while Nix, CodeQL, Dependency
Review and review all completed green on it. gh run rerun and a
close/reopen of the PR produced no new runs, and the API reported the
runs as both completed and queued depending on which endpoint was
asked. Empty commit to get fresh runs scheduled.


---------

+++ src/main/java/org/riptide/flows/parser/sflow/SflowFlowBuilder.java
@@ -133,7 +133,7 @@ public final class SflowFlowBuilder {
             // Scaling by the raw wire rate while reporting a guarded one would produce a row that
             // reads as real unsampled traffic of zero volume — a rate of 0 gives bytes = 0 and
             // packets = 0, with nothing marking it as junk.
-            /**
+            /*
              * The stated frame length, scaled by the stated rate — or nothing, if the frame length is
              * not one a frame could have.
              *
@@ -176,9 +176,9 @@ public final class SflowFlowBuilder {
                 return scale();
             }
 
-            /** The rate the counters are scaled by: the exporter's, or 1 when it stated nothing usable. */
+            /* The rate the counters are scaled by: the exporter's, or 1 when it stated nothing usable. */
             private long scale() {
-                return usable(sample.samplingRate) ? sample.samplingRate : 1L;
+                return usable((double) sample.samplingRate) ? sample.samplingRate : 1L;
             }
 
             @Override
@@ -233,7 +233,7 @@ public final class SflowFlowBuilder {
 
             @Override
             public double getSamplingInterval() {
-                return usable(sample.samplingRate) ? sample.samplingRate : 1.0d;
+                return usable((double) sample.samplingRate) ? (double) sample.samplingRate : 1.0d;
             }
 
             /*
@@ -247,7 +247,7 @@ public final class SflowFlowBuilder {
              */
             @Override
             public SamplingProvenance getSamplingProvenance() {
-                return usable(sample.samplingRate) ? SamplingProvenance.Record : SamplingProvenance.Assumed;
+                return usable((double) sample.samplingRate) ? SamplingProvenance.Record : SamplingProvenance.Assumed;
             }
         };
     }
